+++ stata_kernel/__init__.py
@@ -1,5 +1,5 @@
 """An example Jupyter kernel"""
 
-__version__ = '1.4.7'
+__version__ = '1.6.0'
 
 from .kernel import StataKernel